backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validation and Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
cryptography==42.0.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development
black==23.12.1
ruff==0.1.11
mypy==1.8.0
